Add basic toOmegaString/ParseRelation testing
Includes implementing UF declarations in toOmegaString
